Refactor test-util and parser-tests.
Above all, I defined a directory hierarchy and terminologies for a test collection as follows:

[base]
   |- [Set 1]
         |- test_X.fl
         |- test_Y.fl
         |-    ...
   |- [Set 2]
   |- [Set ..]
   ..
   |- [Set N]

A test collection consists of multiple test sets, each of which contains
multiple test units. All test units in the same test set shares the test ways,
including how to assert an actual result against an the expected result,
how to handle input data. In most cases, an input data is a source code.
A file name in a test set is used as a test name.

I designed TestDriver and TestSet. With TestSet trait, I implemented
a separation between default implementation and some overridable features.