Moved unit tests in parser.rs to parser-tests.
parser-tests is an integration test for parser. It transforms
each source into ast, and generate a source code from the ast.
It verifies the parser and ast generated from the source code exactly.

Adding a test to parser-tests requires only adding a source file and
a expected result. So, it is a very scalable test environment.
I added it for lots of unit tests to be added continuously.

I also added run-tests.sh, which easily launchs all unit tests.
You can execute it as follows:

```
./run-tests
```

If you want to specify some tests, you can use '--filter' option as follows:

```
./run-tests --filter=tuple
```

If so, only tests whose names are matched to a given filter keyword would
be launched.